requirements.txt
pandas
jinja2
weasyprint
PyYAML
requests
schedule